+++ cmd/hidden_lake/service/pkg/client/builder.go
@@ -34,6 +34,6 @@ func (p *sBuilder) Friend(pAliasName string, pPubKey asymmetric.IPubKey) *pkg_se
 func (p *sBuilder) Request(pReceiver string, pReq request.IRequest) *pkg_settings.SRequest {
 	return &pkg_settings.SRequest{
 		FReceiver: pReceiver,
-		FReqData:  pReq.ToString(),
+		FReqData:  pReq.(*request.SRequest),
 	}
 }