+++ examples/echo_service/_request/request.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-JSON_DATA='{
+PUSH_FORMAT='{
+    "receiver":"Bob",
+    "req_data":{
         "method":"POST",
         "host":"hidden-echo-service",
         "path":"/echo",
@@ -8,14 +10,7 @@ JSON_DATA='{
                 "Accept": "application/json"
         },
         "body":"aGVsbG8sIHdvcmxkIQ=="
-}';
-
-JSON_DATA=${JSON_DATA//\"/\\\"} # "method" -> \"method\", ...
-JSON_DATA=${JSON_DATA//[$'\t\r\n ']} # delete ['\t', '\r', '\n', ' '] from string
-
-PUSH_FORMAT='{
-        "receiver":"Bob",
-        "req_data":"'$JSON_DATA'"
+    }
 }';
 
 d="$(date +%s)";